Refactor: Moved `models.py` to `backend/data/` for better organization of database-related files
